Split off Dictionary and Predicate, move compareObjects to Dictionary.compare
