+++ src/runtime/manifest.ts
@@ -16,7 +16,8 @@ import {Id, IdGenerator} from './id.js';
 import {InterfaceInfo} from './interface-info.js';
 import {Handle as InterfaceInfoHandle} from './interface-info.js';
 import {Slot as InterfaceInfoSlot} from './interface-info.js';
-import {Dictionary, Runnable} from '../common/hot.js';
+import {Dictionary} from '../common/collect/dictionary.js';
+import {Runnable} from '../common/hot.js';
 import {Loader} from './loader.js';
 import {ManifestMeta} from './manifest-meta.js';
 import * as AstNode from './manifest-ast-nodes.js';
