+++ src/runtime/storage/crdt-collection-model.ts
@@ -9,7 +9,7 @@
  */
 
 import {assert} from '../../platform/assert-web.js';
-import {Dictionary} from '../../common/hot.js';
+import {Dictionary} from '../../common/collect/dictionary.js';
 
 // Bulding block for CRDT collections. Tracks the membership (keys) of
 // values identified by unique IDs. A value is considered to be part
