+++ src/runtime/storage/firebase/firebase-storage.ts
@@ -13,7 +13,8 @@ import {atob} from '../../../platform/atob-web.js';
 import {btoa} from '../../../platform/btoa-web.js';
 import {firebase} from '../../../platform/firebase-web.js';
 import {Id} from '../../id.js';
-import {Dictionary, Runnable} from '../../../common/hot.js';
+import {Runnable} from '../../../common/hot.js';
+import {Dictionary} from '../../../common/collect/dictionary.js';
 import {BigCollectionType, CollectionType, ReferenceType, Type, TypeVariable} from '../../type.js';
 import {setDiff} from '../../util.js';
 import {CrdtCollectionModel, ModelValue, SerializedModelEntry} from '../crdt-collection-model.js';
